api calls seperated into hooks
